requirements.txt
torch
pytorch-lightning
transformers
datasets
x_transformers
Pillow
numpy
tensorboard

